requirements.txt
Django==4.0.3
djangorestframework==3.13.1
drf-spectacular==0.22.0
psycopg2-binary==2.8.6
django-filter==21.1